Add a Code of Conduct to the repo
Codes of Conduct are vital to any kind of project. Mostly in that it
signals to those margianalized in society that they too have an
equal place at the table and that attempts to make that not happen are
unwelcome. I often think of this bit of explanation from Graydon, the
creator of Rust, about why they added a CoC early on in the project's
history:

'Well, I wrote the initial CoC and put the "We will exclude you from
interaction" phrase in there, so maybe I'll mention the impetus and
meaning.

I was given the opportunity to start a language project by my employer,
Mozilla corp. I'd had the experience of working -- both professionally
and on volunteer time -- with many PL communities in the past.
Communities that were prone to several norms of discourse that I found
extremely difficult to deal with, that would have prevented me, and
several people I knew and wanted to work with, from bothering to work on
a language at all. In other words: I would not have built the language,
nor participated in a project of building the language, if I had to
subject myself to the kind of discourse normally surrounding
language-building communities.

In other other words: the norms of other communities were already
excluding me.