Add initial code, LICENSE, CONTRIBUTING and README
This sets up all of the needed infra that a structure needs like
a LICENSE, CONTRIBUTING, and README file, but the code itself sets the
standard for the repo. All this does is add the `git init` subcommand.
However, this command has tests, docs, and has it's own separate file,
that gets publically reexported in the crate root. Going forward this is
how the repo should be organized so as to separate things by command and
make it easier to navigate to those new to the codebase.